Fixed collisions for shared tooltips.
Stored the DOM node of the caller in tip.data('pTooltip'), and use this
to ensure that the tip is closed only when the right closeTip() function
fires. (Causes issues when multiple elements share the same tooltip.)

Also moved mouseEntered to a local variable, and not a data() of tip.