Used timestamps for collision detection.
The delayed close event checks to see that the tip has not been reshown
in the meantime. (It does this by comparing a timestamp "uid".)

This fixes the following problem: Hover over <a> show tip <t>. Now leave <a>
and quickly return to <a>. Tip closes when it should not.

(Problem is more pronounced when multiple elements show the same tip.)